src/generate/requirements.txt
google-cloud-aiplatform
google-auth
datasets
aiohttp
asyncio